+++ types/signable.go
@@ -1,14 +1,18 @@
 package types
 
 import (
+	"github.com/cometbft/cometbft/crypto/bls12381"
 	"github.com/cometbft/cometbft/crypto/ed25519"
 	cmtmath "github.com/cometbft/cometbft/libs/math"
 )
 
 // MaxSignatureSize is a maximum allowed signature size for the Proposal
 // and Vote.
-// XXX: secp256k1 does not have Size nor MaxSize defined.
-var MaxSignatureSize = cmtmath.MaxInt(ed25519.SignatureSize, 64)
+// XXX: secp256k1 does not have max signature size defined.
+var MaxSignatureSize = cmtmath.MaxInt(
+	ed25519.SignatureSize,
+	bls12381.SignatureLength,
+)
 
 // Signable is an interface for all signable things.
 // It typically removes signatures before serializing.